move zoom files to zoom dir
