+++ apps/router-e2e/__e2e__/link-preview/app/misc.tsx
@@ -159,10 +159,10 @@ const HomeIndex = () => {
             </Pressable>
           </Link.Trigger>
         </Link>
-        <Link href="/zoom-dest" unstable_transition="zoom" unstable_customTransitionSource asChild>
+        <Link href="/zoom-dest" unstable_transition="zoom" asChild>
           <Link.Trigger>
             <Pressable style={{ flex: 3, alignItems: 'center' }}>
-              <Link.TransitionSource>
+              <Link.AppleZoom>
                 <View style={{ width: '100%', aspectRatio: width / height }}>
                   <Image
                     source={require('../../../assets/frog.jpg')}
@@ -170,7 +170,7 @@ const HomeIndex = () => {
                     style={{ width: '100%', height: '100%' }}
                   />
                 </View>
-              </Link.TransitionSource>
+              </Link.AppleZoom>
               <Text style={{ fontSize: 18, fontWeight: 'bold', textAlign: 'center' }}>
                 The frog
               </Text>
