+++ packages/expo-router/src/link/zoom/link-apple-zoom.tsx
@@ -2,7 +2,7 @@
 
 import { Children, use, type PropsWithChildren } from 'react';
 
-import { isZoomTransitionEnabled } from '../ZoomTransitionEnabler.ios';
+import { isZoomTransitionEnabled } from './ZoomTransitionEnabler.ios';
 import { ZoomTransitionSourceContext } from './zoom-transition-context';
 import { LinkZoomTransitionSource } from '../preview/native';
 
@@ -31,7 +31,6 @@ export function LinkAppleZoom({ children, alignmentRect }: LinkAppleZoomProps) {
     );
   }
   const { identifier } = value;
-  console.log('Link.ZoomTransitionSourceWrapper rendering with identifier:', identifier);
   if (Children.count(children) > 1) {
     console.warn(
       '[expo-router] Link.ZoomTransitionSource only accepts a single child component. Please wrap multiple children in a View or another container component.'
