+++ packages/expo-router/src/useScreens.tsx
@@ -13,7 +13,7 @@ import React, { useEffect } from 'react';
 import { LoadedRoute, Route, RouteNode, sortRoutesWithInitial, useRouteNode } from './Route';
 import { useExpoRouterStore } from './global-state/storeContext';
 import EXPO_ROUTER_IMPORT_MODE from './import-mode';
-import { ZoomTransitionEnabler } from './link/ZoomTransitionEnabler';
+import { ZoomTransitionEnabler } from './link/zoom/ZoomTransitionEnabler';
 import {
   hasParam,
   INTERNAL_EXPO_ROUTER_NO_ANIMATION_PARAM_NAME,
